Refactor name to default instead the same as interface
