+++ plugin/src/main/java/com/craftmend/openaudiomc/generic/core/interfaces/OpenAudioInvoker.java
@@ -1,7 +1,6 @@
 package com.craftmend.openaudiomc.generic.core.interfaces;
 
-import com.craftmend.openaudiomc.OpenAudioMc;
-import com.craftmend.openaudiomc.generic.networking.interfaces.INetworkingService;
+import com.craftmend.openaudiomc.generic.networking.interfaces.NetworkingService;
 import com.craftmend.openaudiomc.generic.platform.Platform;
 
 public interface OpenAudioInvoker {
@@ -9,7 +8,7 @@ public interface OpenAudioInvoker {
     boolean hasPlayersOnline();
     boolean isSlave();
     Platform getPlatform();
-    Class<? extends INetworkingService> getServiceClass();
+    Class<? extends NetworkingService> getServiceClass();
     ITaskProvider getTaskProvider();
     ConfigurationImplementation getConfigurationProvider();
