+++ plugin/src/main/java/com/craftmend/openaudiomc/spigot/OpenAudioMcSpigot.java
@@ -5,7 +5,7 @@ import com.craftmend.openaudiomc.generic.core.interfaces.ConfigurationImplementa
 import com.craftmend.openaudiomc.generic.core.interfaces.ITaskProvider;
 import com.craftmend.openaudiomc.generic.core.interfaces.OpenAudioInvoker;
 import com.craftmend.openaudiomc.generic.core.logging.OpenAudioLogger;
-import com.craftmend.openaudiomc.generic.networking.interfaces.INetworkingService;
+import com.craftmend.openaudiomc.generic.networking.interfaces.NetworkingService;
 import com.craftmend.openaudiomc.generic.platform.Platform;
 import com.craftmend.openaudiomc.generic.state.states.WorkerState;
 import com.craftmend.openaudiomc.spigot.modules.commands.SpigotCommandModule;
@@ -156,7 +156,7 @@ public final class OpenAudioMcSpigot extends JavaPlugin implements OpenAudioInvo
     }
 
     @Override
-    public Class<? extends INetworkingService> getServiceClass() {
+    public Class<? extends NetworkingService> getServiceClass() {
         return proxyModule.getMode().getServiceClass();
     }
